+++ docs/utils/check-parent-in-children.ts
@@ -1,0 +1,6 @@
+// Types
+import type { LinkItemProps } from '@/types/components/atoms/list-item'
+
+const CheckParentInChildren = (items: LinkItemProps[], parent?: string) => items.some(item => (!parent && !item.parent) || item.parent === parent)
+
+export default CheckParentInChildren