+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Release Notes
 All notable changes to this package will be documented in this file.
 
+# 0.5.2
 - Created a `CheckParentInChildren` utility
 - Created a back button in the menu **AppBar**